Add Back-end

+++ Project3mon/app/src/main/java/com/example/project3mon/UserAdapter.java
@@ -52,7 +52,7 @@ public class UserAdapter extends RecyclerView.Adapter<UserAdapter.UserViewHolder
         String shortName = name[name.length - 2] + " " + name[name.length - 1];
 
         holder.txtName.setText(shortName);
-        holder.txtPrice.setText(user.getPrice()+ " VNĐ");
+        holder.txtPrice.setText((int)user.getPrice()+ " VNĐ");
 
         holder.layoutItem.setOnClickListener(new View.OnClickListener() {
             @Override
